feat: 깃허브 information data class 추가
